Enhancement: new functionalities (command line arguments + retrieve cache values)
Little code refactor
Added command line parameters (with help and parameters explaination)
Database index now can be specified
Possibility to retrieve a list of values based on a list of cache keys
Simple value transformation for retrieved values (ie for dates, we can retrieve automatically the elapsed minutes)
None of the previouse features has been removed, I simply put that under a specific command type.
