+++ Enums/MonitorCacheValueTransformEnum.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PRTG_Redis_Sensor.Enums
+{
+	public enum MonitorCacheValueTransformEnum
+	{
+		ElapsedMinutes
+	}
+}
