+++ Enums/MonitorServiceEnum.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PRTG_Redis_Sensor.Enums
+{
+	public enum MonitorServiceEnum
+	{
+		Stats,
+		CacheKeysValues
+	}
+}
