+++ Model/PRTGMonitor.cs
@@ -1,0 +1,11 @@
+﻿using Newtonsoft.Json;
+
+namespace PRTG_Redis_Sensor.Model
+{
+	public class PRTGMonitor
+	{
+
+		[JsonProperty(propertyName: "prtg")]
+		public PRTGResponse Response { get; set; }
+	}
+}
